Quote table ticket

+++ springboot/trading-app/src/main/java/ca/jrvs/apps/trading/Model/Domain/IexQuote.java
@@ -384,7 +384,7 @@ public class IexQuote {
         this.lastTradeTime = lastTradeTime;
     }
 
-    public int getLatestPrice() {
+    public double getLatestPrice() {
         return latestPrice;
     }
